requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
loguru==0.7.2
requests>=2.0.0
pydantic>=2.0.0
python-multipart==0.0.6
aiohttp==3.9.3
httpx==0.25.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
cryptography==42.0.2
python-jose[cryptography]==3.3.0
pyyaml>=6.0.0
python-benedict>=0.33.0
typing-extensions>=4.0.0
tweepy>=4.0.0
stripe>=7.11.0
six>=1.16.0
google-api-python-client>=2.0.0
hubspot-api-client>=8.1.0
setuptools>=65.0.0
slack_sdk>=3.34.0
linkup-sdk>=0.2.3
agentsjson==0.1.11
